+++ src/concurrentsim.jl
@@ -1,7 +1,10 @@
 export @simlog, nongreedymultilock, spinlock, get_time_tracker
 
 macro simlog(env, msg) # this should be part of @process or @resumable and just convert @info and company
-    return :(@info(@sprintf("t=%.4f @ %05d : %s", now($(esc(env))), active_process($(esc(env))).bev.id, $(esc(msg)))))
+    return :(@info(
+        @sprintf("t=%.4f @ %05d : %s", now($(esc(env))), active_process($(esc(env))).bev.id, $(esc(msg))),
+        _group=LOG_GROUPS.simulation,
+    ))
 end
 
 @resumable function nongreedymultilock(env::Environment, resources)